Reward app #4 (all but no tests)

+++ src/main/java/org/example/properties/Threshold.java
@@ -4,10 +4,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class Threshold {
-    private double level;
+    private BigDecimal level;
     private int points;
 }
